% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_meta_analysis_stan.R
\name{bayesian_biomarker_test_stan}
\alias{bayesian_biomarker_test_stan}
\title{Bayesian biomarker testing using drug response summary statistic estimates with Stan}
\usage{
bayesian_biomarker_test_stan(
  data,
  min_samples = 10,
  min_pos = 1,
  alpha = 1,
  scaling_flag = TRUE,
  iter = 4000,
  thin = 5,
  pars = c("shifted_mean"),
  include = FALSE,
  refresh = -1,
  ...
)
}
\arguments{
\item{data}{Tibble or data frame containing mean and standard deviation
estimates for the summary statistics in columns `stat.mean` and `stat.sd`,
and biomarker status (0 or 1) in column `status`.}

\item{min_samples}{Minimum number of samples needed.}

\item{min_pos}{Minimum number of positive (biomarker status 1) samples needed.}

\item{alpha}{Tuning parameter for the influence of the estimated standard deviations; the effective standard deviation becomes `stat.sd=stat_sd^alpha`. `alpha` should be between 0 and 1.}

\item{scaling_flag}{Whether to scale (standardize) the summary statistics (default: `TRUE`)}

\item{iter}{Number of iterations.}

\item{thin}{A positive integer specifying the period for saving samples.}

\item{pars}{Parameters of interest (see `rstan::sampling`).}

\item{include}{Whether to include or exclude parameters of interest.}

\item{refresh}{How often to refresh the progress bar. The default is to not display the progress bar (`refresh=-1`).}

\item{...}{Arguments passed to `rstan::sampling`.}
}
\value{
A vector of posterior samples for the effect size parameter of the biomarker status.
}
\description{
Bayesian biomarker testing using drug response summary statistic estimates with Stan
}
\examples{
# Run the sampler to get posterior samples of the association between
# the summary statistics and the biomarker, using the Afatinib example
# dataset
beta_samples = bayesian_biomarker_test_stan(BRCA_Afatinib_Response)

# Plot the posterior distribution of the effect size parameters
hist(beta_samples)

}
